+++ crates/config/src/tests.rs
@@ -24,7 +24,7 @@ memospot:
     let parsed_config = Config::init(&partial_yaml_path).unwrap();
 
     assert_eq!(parsed_config.memos.addr, Some("127.0.0.1".to_string()));
-    assert_eq!(parsed_config.memos.port, Some(0));
+    assert_eq!(parsed_config.memos.port, Some(5230));
     assert_eq!(parsed_config.memos.mode, Some("demo".to_string()));
     assert!(parsed_config.memospot.log.enabled.unwrap());
 }